Experimental: custom static attribute - link_latency
Allows to define link latency without having to specify it when calling
the filter function. Usefull when templates with default filters are
defined
and you wish to avoid defining custom filters on each individual
peering.
Instead, 'link latency <int>;' can be set on a per-protocol basis and
referenced
from a filter scope with 'link_latency' attribute.